refactor(orchestration): move subagent_dispatcher to enhanced_orchestration (#6822)
services/ layer should consume orchestration primitives, not define them.
The dispatcher's reflection-pass loop is core orchestration logic, so it
belongs in enhanced_orchestration/. Test imports updated; class behavior
unchanged.
