+++ autobot-backend/enhanced_orchestration/__init__.py
@@ -16,10 +16,12 @@ Sub-modules:
 - success_criteria.py: Structured success criteria evaluation
 - collaboration_coordinator.py: Redis pub/sub collaboration layer (#6393)
 - agent_router.py: Agent selection, resolution, capability coverage (#6393/#6392)
+- subagent_dispatcher.py: Autonomous subagent spawning for parallel workstreams (#6822)
 """
 
 from .agent_router import AgentRouter
 from .collaboration_coordinator import CollaborationCoordinator
+from .subagent_dispatcher import SubagentDispatcher
 from .execution_strategies import ExecutionStrategyHandler
 from .success_criteria import (
     CriteriaResult,
@@ -65,4 +67,6 @@ __all__ = [
     # Collaborators extracted from WorkflowRunner (#6393/#6392)
     "AgentRouter",
     "CollaborationCoordinator",
+    # Subagent dispatcher relocated from services/orchestration (#6822)
+    "SubagentDispatcher",
 ]
